Add files via upload
Bagels is a deduction game in which the
player tries to guess a random three-digit
number (with no repeating digits) generated
by the computer. After each guess, the
computer gives the player three types of clues:
Bagels None of the three digits guessed is in the secret number.
Pico One of the digits is in the secret number, but the guess has the
digit in the wrong place.
Fermi The guess has a correct digit in the correct place.