Add files via upload
The Sonar Treasure Hunt game in this
chapter is the first to make use of the
Cartesian coordinate system that you
learned about in Chapter 12. This game
also uses data structures, which is just a fancy way
of saying it has list values that contain other lists and
similar complex variables. As the games you program
become more complicated, you’ll need to organize
your data into data structures.