Add files via upload
The program in this chapter isn’t really a
game, but it is fun nevertheless. The program
will convert normal English into a
secret code. It can also convert secret codes
back into regular English. Only someone who knows
the key to the secret codes will be able to understand
the messages.