src code
In this chapter, we’ll make Reversegam,
also known as Reversi or Othello. This
two-player board game is played on a grid,
so we’ll use a Cartesian coordinate system
with x- and y-coordinates. Our version of the game
will have a computer AI that is more advanced than
our Tic-Tac-Toe AI from Chapter 10. In fact, this AI
is so good that it will probably beat you almost every
time you play. (I lose whenever I play against it!)