src code
The Reversegam AI algorithm from
Chapter 15 is simple, but it beats me
almost every time I play it. Because the
computer can process instructions quickly,
it can easily check each possible position on the board
and select the highest-scoring move. It would take me
a long time to find the best move this way.